modified function to print number of chars

+++ holberton.h
@@ -6,6 +6,6 @@
 int _printf(const char *format, ...);
 int _putchar(char c);
 int print_string(char *str);
-int print_num(int n);
+int print_num(long int n);
 
 #endif /* _HOLBERTON_H_ */
